Separate the consolidated dispatcher subscription endpoint
